rename migrations to Capitalize
